+++ datasophon-service/src/main/java/com/datasophon/api/service/ClusterKerberosService.java
@@ -24,9 +24,9 @@ import java.io.IOException;
 
 public interface ClusterKerberosService {
 
-    void downloadUserKeytab(Integer clusterId, String username, HttpServletResponse response) throws IOException;
+    void downloadUserKeytab(Long clusterId, String username, HttpServletResponse response) throws IOException;
 
-    void downloadKeytab(Integer clusterId, String principal, String keytabName, String hostname,
+    void downloadKeytab(Long clusterId, String principal, String keytabName, String hostname,
                         HttpServletResponse response) throws IOException;
 
     void uploadKeytab(MultipartFile file, String hostname, String keytabFileName) throws IOException;
